Fixed problem with context variables.
Now the context is a deBruijn indexed list of the
type of the variables.

We cannot use directly the function offered in the reflection
API `getContext` because the type of each variable refers to
the local context where such variable is typed. We instead require
that the types are global in the context of the function.

Still needs a small refactor.